# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: VertexLearn AI API
  version: 0.7.0
  description: Phase 7 surface — auth, course foundation, learning core, AI tutor/RAG (course-isolated), summaries, AI quiz drafts, flashcards, study plans, mastery, recommendations, admin workspace, course approvals, discussions, announcements, notifications, email abstraction, S3/MinIO storage (uploads, signed URLs), catalog filters/reviews, payments/revenue, Redis caching (catalog/course/leaderboard, x-cache), rate limits (auth 10/min/IP, AI 20/min/user, global 100/min), gamification leaderboard, health/ready aliases.
servers:
  - url: http://localhost:4000
paths:
...
  /api/v1/assignments/{id}/upload:
    post:
      summary: Binary assignment upload (multipart file field; PDF/ZIP/code-text with magic-byte checks; bytes to object storage, metadata in Postgres)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Submitted with storage backend metadata }
        '400': { description: Missing file, type not accepted, or over size limit }
        '403': { description: Enrollment required or deadline passed }
        '404': { description: Assignment not found }
        '409': { description: Already submitted }
  /api/v1/submissions/{id}/download:
    get:
      summary: Download a submission's file (owning student, assignment/course owner, or admin; objects stay private)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: File bytes }
        '403': { description: Not authorized }
        '404': { description: Submission or file not found }
  /api/v1/lectures/{id}/upload-url:
    post:
      summary: Mint a presigned PUT for a lecture video asset (own course instructor or admin; stores video_key)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Upload target with presigned URL }
        '403': { description: Not the course owner }
        '404': { description: Lecture not found }
  /api/v1/lectures/{id}/video-url:
    get:
      summary: Access-checked signed playback URL (enrolled/owner/admin; honest unavailable state without storage backend; production HLS/CDN out of scope)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Signed URL or unavailable state }
        '403': { description: Enrollment required }
        '404': { description: Lecture not found or no video attached }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT